Restructuring: moved utility functions in __init__.py to utils.py; moved User class definition in __init__.py to user.py; moved webauthn routes from backend to fido_app; adjusted imports; fixed some typos
